% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdbg.R
\name{pdbg}
\alias{pdbg}
\title{debug pipe chain}
\usage{
pdbg(.expr)
}
\arguments{
\item{expr}{expressions to go through, if not provided content of clipboard will be used}
}
\description{
Shows the output step by step. press \emph{s} to skip (i.e. not print, it still evaluates),
\emph{q} to quit. Explore environment as in \code{base::debug}, current output is stored in
\code{.} variable.
}
\examples{
\dontrun{
library(dplyr)
debug_pipe(
test <- iris \%>\%
 slice(1:2) \%>\%
 select(1:3) \%>\%
 mutate(x=3)
)

# works in functions too :
fun <- function(){
 print("hello")
 debug_pipe(
   test <- iris \%>\%
     slice(1:2) \%>\%
     select(1:3) \%>\%
     mutate(x=3)
 )
 print("bye")
}

fun()
}
}
